requirements.txt
openai-whisper==20231117
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.3
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
